Installed react-router-dom and updated some components
Doesn't quite work as expected. When I click one of the `Link` that is imported from the package, it changes the url in the address bar but doesn't update which component is rendered.

This may be a scoping issue, as I have used the `Link` component in the `Nav` component that I created and I'm using the `BrowserRouter`, `Switch`, and `Route`s in the newly made `MainContainer` component.

Will address this in follow up PR